# contrato/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JEJE",
    "version": "0.1.0"
  },
  "paths": {
...
    "/minhas/transacoes": {
      "get": {
        "summary": "Minhas Transacoes",
        "operationId": "minhas_transacoes_minhas_transacoes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transacao"
                  },
                  "title": "Response Minhas Transacoes Minhas Transacoes Get"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente, inválida ou expirada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minhas/transacoes/{transaction_id}": {
      "get": {
        "summary": "Minha Transacao",
        "operationId": "minha_transacao_minhas_transacoes__transaction_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9-]{1,64}$",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transacao"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente, inválida ou expirada"
          },
          "404": {
            "description": "Transação não encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minhas/transacoes/{transaction_id}/situacao": {
      "get": {
        "summary": "Situacao",
        "description": "Fatos da transação e a decisão da política para uma consulta sobre ela (POL-CON-*).",
        "operationId": "situacao_minhas_transacoes__transaction_id__situacao_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9-]{1,64}$",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Situacao"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente, inválida ou expirada"
          },
          "404": {
            "description": "Transação não encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minhas/transacoes/{transaction_id}/contestacao": {
      "get": {
        "summary": "Avaliar Contestacao",
        "description": "Avalia, sem criar nada, se uma contestação desta transação pode virar pré-caso.",
        "operationId": "avaliar_contestacao_minhas_transacoes__transaction_id__contestacao_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9-]{1,64}$",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisaoDaPolitica"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente, inválida ou expirada"
          },
          "404": {
            "description": "Transação não encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minhas/transacoes/{transaction_id}/contestacao/proposta": {
      "post": {
        "summary": "Propor",
        "description": "Avalia a contestação e, se a política permitir, cria uma proposta a confirmar.",
        "operationId": "propor_minhas_transacoes__transaction_id__contestacao_proposta_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9-]{1,64}$",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoDeContestacao"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente, inválida ou expirada"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoDeContestacao"
                }
              }
            },
            "description": "Created"
          },
          "404": {
            "description": "Transação não encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minhas/propostas/{proposta_id}/confirmacao": {
      "post": {
        "summary": "Confirmar",
        "description": "Confirma a proposta: grava o pré-caso sem duplicar e só responde depois de relê-lo.",
        "operationId": "confirmar_minhas_propostas__proposta_id__confirmacao_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Proposta Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreCaso"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente, inválida ou expirada"
          },
          "200": {
            "description": "Já confirmado antes: mesmo protocolo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreCaso"
                }
              }
            }
          },
          "404": {
            "description": "Proposta não encontrada"
          },
          "409": {
            "description": "Proposta vencida ou situação da transação mudou"
          },
          "503": {
            "description": "Pré-caso não registrado; nada foi criado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/minhas/pre-casos": {
      "get": {
        "summary": "Meus Pre Casos",
        "description": "Acompanhamento: pré-casos do cliente da sessão, mais recentes primeiro.",
        "operationId": "meus_pre_casos_minhas_pre_casos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PreCaso"
                  },
                  "type": "array",
                  "title": "Response Meus Pre Casos Minhas Pre Casos Get"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente, inválida ou expirada"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AvaliacaoDeContestacao": {
        "properties": {
          "decisao": {
            "$ref": "#/components/schemas/DecisaoDaPolitica"
          },
          "proposta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Proposta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "decisao",
          "proposta"
        ],
        "title": "AvaliacaoDeContestacao"
      },
...
      "DecisaoDaPolitica": {
        "properties": {
          "regra": {
            "type": "string",
            "title": "Regra"
          },
          "acao": {
            "type": "string",
            "title": "Acao"
          },
          "detalhe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalhe"
          }
        },
        "type": "object",
        "required": [
          "regra",
          "acao",
          "detalhe"
        ],
        "title": "DecisaoDaPolitica"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreCaso": {
        "properties": {
          "protocolo": {
            "type": "string",
            "title": "Protocolo"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "type": "object",
        "required": [
          "protocolo",
          "transaction_id",
          "estado",
          "criado_em"
        ],
        "title": "PreCaso"
      },
      "Proposta": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "expira_em": {
            "type": "string",
            "format": "date-time",
            "title": "Expira Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "transaction_id",
          "expira_em"
        ],
        "title": "Proposta"
      },
...
      "Situacao": {
        "properties": {
          "transacao": {
            "$ref": "#/components/schemas/Transacao"
          },
          "decisao": {
            "$ref": "#/components/schemas/DecisaoDaPolitica"
          }
        },
        "type": "object",
        "required": [
          "transacao",
          "decisao"
        ],
        "title": "Situacao"
      },
      "Transacao": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "transaction_date": {
            "type": "string",
            "format": "date-time",
            "title": "Transaction Date"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "transaction_status": {
            "type": "string",
            "title": "Transaction Status"
          },
          "response_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Code"
          },
          "transaction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Type"
          },
          "merchant_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant Name"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "transaction_date",
          "amount",
          "currency",
          "transaction_status",
          "response_code",
          "transaction_type",
          "merchant_name",
          "channel"
        ],
        "title": "Transacao"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Token da sessão de teste (POST /sessoes)",
        "scheme": "bearer"
      }
    }
  }
}